Retrieve Result Object by ID

This function loads a serialized result object from the internal package directory based on a given ID. The result object is expected to be stored as an RDS file named ResObj_<id>.rds within the res folder of the USERM package.

Usage

getRes(id, custom_dir = NULL)

Arguments

id

A character string representing the ID of the result object to retrieve.

Value

An R object containing the result data associated with the specified ID.

Details

The function constructs the file path using system.file() and checks whether the corresponding RDS file exists. If the file is found, it is read and returned. If not, the function throws an error indicating that the result object could not be found.

Examples

if (FALSE) { # \dontrun{
  res <- getRes("SCC_Bead_CD4_NFR700")
  print(res)
} # }