Visualize the Intercept Matrix from a Result Object

This function creates a heatmap visualization of the interceptMtx matrix contained within a result object Res. It uses a customizable color gradient and overlays numeric values on each cell.

Usage

checkRes_interceptMtx(
  Res,
  min = -10,
  mid = 0,
  max = 10,
  mincolor = "blue",
  midcolor = "white",
  maxcolor = "red"
)

Arguments

Res

A Res object.

min

Numeric value representing the minimum of the color scale. Default is -10.

mid

Numeric value representing the midpoint of the color scale. Default is 0.

max

Numeric value representing the maximum of the color scale. Default is 10.

mincolor

Color corresponding to the minimum value. Default is "blue".

midcolor

Color corresponding to the midpoint value. Default is "white".

maxcolor

Color corresponding to the maximum value. Default is "red".

Value

A Heatmap object from the ComplexHeatmap package, visualizing the interceptMtx with annotated values.

Examples

if (FALSE) { # \dontrun{
  checkRes_interceptMtx(result)
} # }